model_worker/requirements.txt
opencv-python-headless
redis
pillow
numpy
ultralytics
torch
psycopg2-binary
